feat(osd-viewer-demo-v2): code optimizations
