+++ apps/viewer-demo-v2/src/hooks/useWebGL/func.ts
@@ -1,0 +1,72 @@
+import { FragmentShader, VertexShader } from './types'
+
+export function createShader(
+  gl: WebGLRenderingContext,
+  type: VertexShader | FragmentShader,
+  source: string
+) {
+  const shader = gl.createShader(type)
+  if (!shader) {
+    console.error('Failed to create shader')
+    return
+  }
+  gl.shaderSource(shader, source)
+  gl.compileShader(shader)
+  const success = gl.getShaderParameter(shader, gl.COMPILE_STATUS)
+  if (success) {
+    return shader
+  }
+  gl.deleteShader(shader)
+}
+
+export function createProgram(
+  gl: WebGLRenderingContext,
+  vertexShader: WebGLShader,
+  fragmentShader: WebGLShader
+) {
+  const program = gl.createProgram()
+  if (!program) {
+    console.error('Failed to create program')
+    return
+  }
+
+  gl.attachShader(program, vertexShader)
+  gl.attachShader(program, fragmentShader)
+  gl.linkProgram(program)
+  const success = gl.getProgramParameter(program, gl.LINK_STATUS)
+  if (success) {
+    return program
+  }
+  gl.deleteProgram(program)
+}
+
+export function setProgram(
+  gl: WebGLRenderingContext,
+  source: { vertex: string; fragment: string }
+) {
+  const vertexShader = createShader(gl, gl.VERTEX_SHADER, source.vertex)
+  const fragmentShader = createShader(gl, gl.FRAGMENT_SHADER, source.fragment)
+  if (!vertexShader || !fragmentShader) return
+  return createProgram(gl, vertexShader, fragmentShader)
+}
+
+export function initializeWebGL(
+  gl: WebGLRenderingContext,
+  source: { vertex: string; fragment: string },
+  vertexData: number[]
+) {
+  const program = setProgram(gl, source)
+
+  if (!program) return
+
+  const positionAttributeLocation = gl.getAttribLocation(program, 'a_position')
+  const positionBuffer = gl.createBuffer()
+  gl.bindBuffer(gl.ARRAY_BUFFER, positionBuffer)
+  gl.bufferData(gl.ARRAY_BUFFER, new Float32Array(vertexData), gl.STATIC_DRAW)
+
+  return {
+    program,
+    positionAttributeLocation,
+    positionBuffer,
+  }
+}
