+++ apps/viewer-demo-v2/src/hooks/useWebGL/types.ts
@@ -1,0 +1,8 @@
+export type VertexShader = WebGL2RenderingContext['VERTEX_SHADER']
+export type FragmentShader = WebGL2RenderingContext['FRAGMENT_SHADER']
+
+export interface GLConfig {
+  program: WebGLProgram
+  positionAttributeLocation: number
+  positionBuffer: WebGLBuffer | null
+}
